+++ src/App.js
@@ -31,30 +31,30 @@ import AcuerdosINE2023 from "./pages/acuerdos/acuerdosINE/AcuerdosINE2023";
 
 //Acuerdos ITE
 import AcuerdosAnteriores from "./pages/acuerdos/AcuerdosAnteriores";
-import Acuerdos1998 from "./pages/acuerdos/Acuerdos1998";
-import Acuerdos2001 from "./pages/acuerdos/Acuerdos2001";
-import Acuerdos2002 from "./pages/acuerdos/Acuerdos2002";
-import Acuerdos2003 from "./pages/acuerdos/Acuerdos2003";
-import Acuerdos2004 from "./pages/acuerdos/Acuerdos2004";
-import Acuerdos2005 from "./pages/acuerdos/Acuerdos2005";
-import Acuerdos2006 from "./pages/acuerdos/Acuerdos2006";
-import Acuerdos2007 from "./pages/acuerdos/Acuerdos2007";
-import Acuerdos2008 from "./pages/acuerdos/Acuerdos2008";
-import Acuerdos2009 from "./pages/acuerdos/Acuerdos2009";
-import Acuerdos2010 from "./pages/acuerdos/Acuerdos2010";
-import Acuerdos2011 from "./pages/acuerdos/Acuerdos2011";
-import Acuerdos2012 from "./pages/acuerdos/Acuerdos2012";
-import Acuerdos2013 from "./pages/acuerdos/Acuerdos2013";
-import Acuerdos2014 from "./pages/acuerdos/Acuerdos2014";
-import Acuerdos2015 from "./pages/acuerdos/Acuerdos2015";
-import Acuerdos2016 from "./pages/acuerdos/Acuerdos2016";
-import Acuerdos2017 from "./pages/acuerdos/Acuerdos2017";
-import Acuerdos2018 from "./pages/acuerdos/Acuerdos2018";
-import Acuerdos2019 from "./pages/acuerdos/Acuerdos2019";
-import Acuerdos2020 from "./pages/acuerdos/Acuerdos2020";
-import Acuerdos2021 from "./pages/acuerdos/Acuerdos2021";
-import Acuerdos2022 from "./pages/acuerdos/Acuerdos2022";
-import Acuerdos2023 from "./pages/acuerdos/Acuerdos2023";
+import Acuerdos1998 from "./pages/acuerdos/acuerdosITE/Acuerdos1998";
+import Acuerdos2001 from "./pages/acuerdos/acuerdosITE/Acuerdos2001";
+import Acuerdos2002 from "./pages/acuerdos/acuerdosITE/Acuerdos2002";
+import Acuerdos2003 from "./pages/acuerdos/acuerdosITE/Acuerdos2003";
+import Acuerdos2004 from "./pages/acuerdos/acuerdosITE/Acuerdos2004";
+import Acuerdos2005 from "./pages/acuerdos/acuerdosITE/Acuerdos2005";
+import Acuerdos2006 from "./pages/acuerdos/acuerdosITE/Acuerdos2006";
+import Acuerdos2007 from "./pages/acuerdos/acuerdosITE/Acuerdos2007";
+import Acuerdos2008 from "./pages/acuerdos/acuerdosITE/Acuerdos2008";
+import Acuerdos2009 from "./pages/acuerdos/acuerdosITE/Acuerdos2009";
+import Acuerdos2010 from "./pages/acuerdos/acuerdosITE/Acuerdos2010";
+import Acuerdos2011 from "./pages/acuerdos/acuerdosITE/Acuerdos2011";
+import Acuerdos2012 from "./pages/acuerdos/acuerdosITE/Acuerdos2012";
+import Acuerdos2013 from "./pages/acuerdos/acuerdosITE/Acuerdos2013";
+import Acuerdos2014 from "./pages/acuerdos/acuerdosITE/Acuerdos2014";
+import Acuerdos2015 from "./pages/acuerdos/acuerdosITE/Acuerdos2015";
+import Acuerdos2016 from "./pages/acuerdos/acuerdosITE/Acuerdos2016";
+import Acuerdos2017 from "./pages/acuerdos/acuerdosITE/Acuerdos2017";
+import Acuerdos2018 from "./pages/acuerdos/acuerdosITE/Acuerdos2018";
+import Acuerdos2019 from "./pages/acuerdos/acuerdosITE/Acuerdos2019";
+import Acuerdos2020 from "./pages/acuerdos/acuerdosITE/Acuerdos2020";
+import Acuerdos2021 from "./pages/acuerdos/acuerdosITE/Acuerdos2021";
+import Acuerdos2022 from "./pages/acuerdos/acuerdosITE/Acuerdos2022";
+import Acuerdos2023 from "./pages/acuerdos/acuerdosITE/Acuerdos2023";
 
 //Demarcaciones
 import Acuerdos from "./pages/demarcaciones/Acuerdos";
