+++ src/pages/acuerdos/Acuerdos.css
@@ -2,12 +2,13 @@
   background: #972069;
   color: #ffffff;
 }
-.AcuerdosAnteriores {
+
+.AcuerdosITE {
   display: grid;
   gap: 20px;
 }
 
-.AcuerdosAnteriores_link:hover {
+.AcuerdosITE_link:hover {
   transform: scale(1.1);
   transition: 0.2s;
 }
@@ -23,9 +24,10 @@
 
 /* general */
 @media only screen and (min-width: 1024px) {
-  .AcuerdosAnteriores {
+  .AcuerdosITE {
     grid-template-columns: auto auto auto auto auto auto;
   }
+
   .AcuerdosINE {
     grid-template-columns: auto auto auto auto;
   }
@@ -33,9 +35,10 @@
 
 /* tablets */
 @media only screen and (min-width: 768px) and (max-width: 1023px) {
-  .AcuerdosAnteriores {
+  .AcuerdosITE {
     grid-template-columns: auto auto auto auto;
   }
+
   .AcuerdosINE {
     grid-template-columns: auto auto auto auto;
   }
@@ -43,9 +46,12 @@
 
 /* mobil */
 @media only screen and (max-width: 767px) {
-  .AcuerdosAnteriores {
+  .AcuerdosITE {
     grid-template-columns: auto auto;
+    /* gap: 50px; */
+    padding: 0 30px;
   }
+
   .AcuerdosINE {
     grid-template-columns: auto auto;
   }
