+++ src/pages/acuerdos/acuerdosITE/Acuerdos2003.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
-import { dataAcuerdos2003 } from "../../assets/data/dataAcuerdos";
+import { dataAcuerdos2003 } from "../../../assets/data/dataAcuerdos";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilePdf } from "@fortawesome/free-solid-svg-icons";
-import top2003 from "../../img/imgA/h_2003.png";
+import top2003 from "../../../img/imgA/h_2003.png";
 
 export default class Acuerdos2003 extends Component {
   render() {
