+++ src/pages/acuerdos/acuerdosITE/Acuerdos2006.js
@@ -1,16 +1,16 @@
 import React, { Component } from "react";
-import { dataAcuerdos2002 } from "../../assets/data/dataAcuerdos";
+import { dataAcuerdos2006 } from "../../../assets/data/dataAcuerdos";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilePdf } from "@fortawesome/free-solid-svg-icons";
-import top2002 from "../../img/imgA/h_2002.png";
+import top2006 from "../../../img/imgA/h_2006.png";
 
-export default class Acuerdos2002 extends Component {
+export default class Acuerdos2006 extends Component {
   render() {
     return (
       <div>
         <img
-          src={top2002}
-          alt="Acuerdos 2002"
+          src={top2006}
+          alt="Acuerdos 2006"
           className="img-fluid w-100 mb-3"
         />
         <table className="table table-hover table-responsive align-middle">
@@ -23,7 +23,7 @@ export default class Acuerdos2002 extends Component {
             </tr>
           </thead>
           <tbody>
-            {dataAcuerdos2002.map((acuerdo) => {
+            {dataAcuerdos2006.map((acuerdo) => {
               return (
                 <tr key={acuerdo.id}>
                   <th scope="row">{acuerdo.id}</th>
