+++ src/pages/acuerdos/acuerdosITE/Acuerdos2007.js
@@ -1,16 +1,16 @@
 import React, { Component } from "react";
-import { dataAcuerdos2004 } from "../../assets/data/dataAcuerdos";
+import { dataAcuerdos2007 } from "../../../assets/data/dataAcuerdos";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilePdf } from "@fortawesome/free-solid-svg-icons";
-import top2004 from "../../img/imgA/h_2004.png";
+import top2007 from "../../../img/imgA/h_2007.png";
 
-export default class Acuerdos2004 extends Component {
+export default class Acuerdos2007 extends Component {
   render() {
     return (
       <div>
         <img
-          src={top2004}
-          alt="Acuerdos 2004"
+          src={top2007}
+          alt="Acuerdos 2007"
           className="img-fluid w-100 mb-3"
         />
         <table className="table table-hover table-responsive align-middle">
@@ -23,7 +23,7 @@ export default class Acuerdos2004 extends Component {
             </tr>
           </thead>
           <tbody>
-            {dataAcuerdos2004.map((acuerdo) => {
+            {dataAcuerdos2007.map((acuerdo) => {
               return (
                 <tr key={acuerdo.id}>
                   <th scope="row">{acuerdo.id}</th>
