+++ src/pages/acuerdos/acuerdosITE/Acuerdos2008.js
@@ -1,16 +1,16 @@
 import React, { Component } from "react";
-import { dataAcuerdos2005 } from "../../assets/data/dataAcuerdos";
+import { dataAcuerdos2008 } from "../../../assets/data/dataAcuerdos";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilePdf } from "@fortawesome/free-solid-svg-icons";
-import top2005 from "../../img/imgA/h_2005.png";
+import top2008 from "../../../img/imgA/h_2008.png";
 
-export default class Acuerdos2005 extends Component {
+export default class Acuerdos2008 extends Component {
   render() {
     return (
       <div>
         <img
-          src={top2005}
-          alt="Acuerdos 2005"
+          src={top2008}
+          alt="Acuerdos 2008"
           className="img-fluid w-100 mb-3"
         />
         <table className="table table-hover table-responsive align-middle">
@@ -23,7 +23,7 @@ export default class Acuerdos2005 extends Component {
             </tr>
           </thead>
           <tbody>
-            {dataAcuerdos2005.map((acuerdo) => {
+            {dataAcuerdos2008.map((acuerdo) => {
               return (
                 <tr key={acuerdo.id}>
                   <th scope="row">{acuerdo.id}</th>
