+++ src/pages/acuerdos/acuerdosITE/Acuerdos2014.js
@@ -1,16 +1,16 @@
 import React, { Component } from "react";
-import { dataAcuerdos2006 } from "../../assets/data/dataAcuerdos";
+import { dataAcuerdos2014 } from "../../../assets/data/dataAcuerdos";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilePdf } from "@fortawesome/free-solid-svg-icons";
-import top2006 from "../../img/imgA/h_2006.png";
+import top2014 from "../../../img/imgA/h_2014.png";
 
-export default class Acuerdos2006 extends Component {
+export default class Acuerdos2014 extends Component {
   render() {
     return (
       <div>
         <img
-          src={top2006}
-          alt="Acuerdos 2006"
+          src={top2014}
+          alt="Acuerdos 2014"
           className="img-fluid w-100 mb-3"
         />
         <table className="table table-hover table-responsive align-middle">
@@ -23,7 +23,7 @@ export default class Acuerdos2006 extends Component {
             </tr>
           </thead>
           <tbody>
-            {dataAcuerdos2006.map((acuerdo) => {
+            {dataAcuerdos2014.map((acuerdo) => {
               return (
                 <tr key={acuerdo.id}>
                   <th scope="row">{acuerdo.id}</th>
